% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{App}
\alias{App}
\title{Generic class for inheritance for app controllers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
App
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{ns}}{\code{\link{ANY}}
App object}
}}
\section{Methods}{


\itemize{
 \item{See superclass} {
   \code{\link[eventr]{IApp}}
 }
 \item{initialize} {

   \itemize{
     \item{ns }{\code{\link{ANY}}}
   }
 }
}
}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}
\keyword{datasets}

