% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{App.Timetracking}
\alias{App.Timetracking}
\title{Connector for Timetracking}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
App.Timetracking
}
\description{
Methods for controlling the app \emph{Timetracking}.
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{ns}}{\code{\link{character}}
App name/class}
}}
\section{Methods}{


\itemize{
 \item{See superclass} {
   \code{\link[eventr]{IApp}}
 }
}
}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}
\keyword{datasets}

