% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateTimeUntilComplete}
\alias{calculateTimeUntilComplete}
\title{Calculate time difference}
\usage{
calculateTimeUntilComplete(table, uid = character())
}
\description{
Calculate time difference.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

