% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculateWeek}
\alias{calculateWeek}
\title{Calculate calendar week}
\usage{
calculateWeek(table, uid = character())
}
\description{
Calculate calendar week.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

