% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deleteData}
\alias{deleteData}
\title{Delete database entry}
\usage{
deleteData(table, uid = character(), dependent = character())
}
\description{
Delete database entry.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

