% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deleteDependent}
\alias{deleteDependent}
\title{Delete dependent}
\usage{
deleteDependent(table, refuid = character())
}
\description{
Delete dependent.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

