% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handleTime}
\alias{handleTime}
\title{Handle time input}
\usage{
handleTime(input, globals = list(hours_per_workday = 8))
}
\description{
Transforms time input in a way that results in a numeric corresponding to
hours.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

