% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{IApp}
\alias{IApp}
\title{Class that functions as an interface for app controllers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IApp
}
\description{
Defines the interface for app controllers.
}
\details{
The terms \emph{interace} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}
\section{Getters/setters}{


\itemize{
}
}

\section{Public methods}{


\itemize{
 \item{importData} {
 }
}
}
\examples{

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}
\keyword{datasets}

