% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadData}
\alias{loadData}
\title{Load from database}
\usage{
loadData(table, uid = character(), refuid = character())
}
\description{
Load from database.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

