% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logTime}
\alias{logTime}
\title{Log time}
\usage{
logTime(db, table, uid = character(), values = list())
}
\description{
log time.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

