% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareDatabase}
\alias{prepareDatabase}
\title{Prepare database}
\usage{
prepareDatabase(sqlite_path, public_fields_compact, public_fields_details,
  private_fields, times_public_fields_compact, times_public_fields_details,
  times_private_fields, projects_public_fields_compact,
  projects_public_fields_details, projects_private_fields, ...)
}
\description{
Prepare database.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

