% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareForDisplay}
\alias{prepareForDisplay}
\title{Prepare DB table for display}
\usage{
prepareForDisplay(input, fields)
}
\description{
Prepare DB table for display.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

