% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saveData}
\alias{saveData}
\title{Save to database}
\usage{
saveData(data, table, uid = character(), globals = list(hours_per_workday =
  8))
}
\description{
Writes stuff to the database in the appropriate format/way.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

