% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saveProject}
\alias{saveProject}
\title{Save project to database}
\usage{
saveProject(data, table, uid = character())
}
\description{
Writes project stuff to the database in the appropriate format/way.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

