% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateSpecific}
\alias{updateSpecific}
\title{Update specific}
\usage{
updateSpecific(table, uid = character(), values, refuid = character())
}
\description{
Update specific.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

