% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VALIDS_issueStatus}
\alias{VALIDS_issueStatus}
\title{Valid issue statuses}
\usage{
VALIDS_issueStatus()
}
\description{
Valid issue statuses.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

