% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validStatuses}
\alias{validStatuses}
\title{Valid statuses}
\usage{
validStatuses()
}
\description{
Valid statuses.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/timetrackr}
}

